move reality-related files to their own folders
