+++ src/core/reality/imaginary-upgrades.js
@@ -1,5 +1,5 @@
-import { BitPurchasableMechanicState, RebuyableMechanicState } from "./game-mechanics";
-import { DC } from "./constants";
+import { BitPurchasableMechanicState, RebuyableMechanicState } from "../game-mechanics";
+import { DC } from "../constants";
 
 class ImaginaryUpgradeState extends BitPurchasableMechanicState {
   constructor(config) {
