+++ src/core/reality/index.js
@@ -1,0 +1,6 @@
+export * from "./reality.js";
+export * from "./reality-upgrades.js";
+export * from "./imaginary-upgrades.js";
+export * from "./perks.js";
+export * from "./black-hole.js";
+export * from "./machines.js";
