+++ src/core/reality/reality.js
@@ -1,4 +1,4 @@
-import { DC } from "./constants";
+import { DC } from "../constants";
 
 /**
  * Object that manages the selection of glyphs offered to the player
